+++ docs/source/api/fabricius.plugins.rst
@@ -4,10 +4,10 @@ fabricius.plugins package
 Submodules
 ----------
 
-fabricius.plugins.generator module
-----------------------------------
+fabricius.plugins.define.generator module
+-----------------------------------------
 
-.. automodule:: fabricius.plugins.generator
+.. automodule:: fabricius.plugins.define.generator
    :members:
    :undoc-members:
    :show-inheritance:
