--- fabricius/generator/__init__.py
@@ -1,1 +1,0 @@
-from .generator import Generator as Generator
