+++ fabricius/plugins/define/__init__.py
@@ -1,0 +1,1 @@
+from .generator import GeneratorPlugin as GeneratorPlugin
