+++ tests/test_plugin.py
@@ -1,33 +1,33 @@
 import unittest
 
-from fabricius.interfaces import BasePlugin
+from fabricius.plugin import Plugin
 
 
-class TestBasePlugin(unittest.TestCase):
+class TestPlugin(unittest.TestCase):
     """
-    Test Fabricius's BasePlugin.
+    Test Fabricius's Plugin.
     """
 
     def test_piid(self):
 
-        class NeutralPlugin(BasePlugin):
+        class NeutralPlugin(Plugin):
             pass
 
         self.assertFalse(NeutralPlugin().has_valid_piid)
         self.assertIsNone(NeutralPlugin().plugin_name)
         self.assertIsNone(NeutralPlugin().plugin_id)
 
-        class TooShortIDPlugin(BasePlugin):
+        class TooShortIDPlugin(Plugin):
             PIID = "abc-12345"
 
         self.assertFalse(TooShortIDPlugin().has_valid_piid)
 
-        class TooShortNamePlugin(BasePlugin):
+        class TooShortNamePlugin(Plugin):
             PIID = "ob-123456"
 
         self.assertFalse(TooShortNamePlugin().has_valid_piid)
 
-        class ValidPlugin(BasePlugin):
+        class ValidPlugin(Plugin):
             PIID = "myplugin-123456"
 
         self.assertTrue(ValidPlugin().has_valid_piid)
